UserSync: Move FetchSyncedUserHook to UserSync
