+++ pkg/services/authn/authnimpl/service.go
@@ -152,7 +152,7 @@ func ProvideService(
 		s.RegisterPostAuthHook(sync.ProvideOauthTokenSync(oauthTokenService, sessionService).SyncOauthToken, 60)
 	}
 
-	s.RegisterPostAuthHook(sync.ProvideFetchUserSync(userService).FetchSyncedUserHook, 100)
+	s.RegisterPostAuthHook(userSyncService.FetchSyncedUserHook, 100)
 
 	return s
 }
