--- pkg/services/authn/authnimpl/sync/fetch_user_sync_test.go
@@ -1,40 +1,0 @@
-package sync
-
-import (
-	"context"
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	"github.com/grafana/grafana/pkg/services/authn"
-)
-
-func TestFetchUserSync_FetchSyncedUserHook(t *testing.T) {
-	type testCase struct {
-		desc        string
-		req         *authn.Request
-		identity    *authn.Identity
-		expectedErr error
-	}
-
-	tests := []testCase{
-		{
-			desc:     "should skip hook when flag is not enabled",
-			req:      &authn.Request{},
-			identity: &authn.Identity{ClientParams: authn.ClientParams{FetchSyncedUser: false}},
-		},
-		{
-			desc:     "should skip hook when identity is not a user",
-			req:      &authn.Request{},
-			identity: &authn.Identity{ID: "apikey:1", ClientParams: authn.ClientParams{FetchSyncedUser: true}},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.desc, func(t *testing.T) {
-			s := ProvideFetchUserSync(nil)
-			err := s.FetchSyncedUserHook(context.Background(), tt.identity, tt.req)
-			require.ErrorIs(t, err, tt.expectedErr)
-		})
-	}
-}
